[BCM2711] - Raspberry Pi 4
Have taken the most recent commit from RPi Foundation repo, and brought it in line with this version.
It appears now to be capable of building the Pi4 w/wo GIC versions. Fires up the Pi4 OK, although
regression tests should now be performed to confirm Pi 3 and 3B+ continue to build/deploy.